feat(View): enable showing errors in SideBar ErrorPanel
move SideBarContext below DocumentEditorContext to share data
remove redundant WizzardEditorTab and refactor structure accordingly
