+++ app/lib/app/SecvisogramPage/View/WizardTab/editors.js
