+++ app/lib/app/SecvisogramPage/View/WizardTab/schema.js
