+++ app/lib/app/SecvisogramPage/View/WizardTab/shared/types.ts
