+++ app/lib/app/SecvisogramPage/View/shared/context/SideBarContext.js
